PinCellCurvy component isolated and applied
