+++ app/components/PinInputCurvy/index.js
@@ -8,167 +8,7 @@ import {
   TouchableWithoutFeedback,
   Alert,
 } from 'react-native';
-
-const PinBox = forwardRef((PinBoxProps, forwardingRef) => {
-  const { size, active, hasValue } = PinBoxProps;
-
-  const checkboxAnim = new Animated.Value(0);
-  const overlayAnim = new Animated.Value(0);
-  const sizeO =
-    typeof size === 'string' ? parseInt(size.replace('%', ''), 10) : size;
-  const heightOverSet = sizeO;
-
-  const [pinBoxStates, setPinBoxStates] = useState({
-    active: !!active,
-    hasValue: !!hasValue,
-    MaxHeight: sizeO + heightOverSet,
-    MaxWidth: sizeO,
-    value: '',
-  });
-
-  const boxStyles = {
-    height: checkboxAnim.interpolate({
-      inputRange: [0, 1],
-      outputRange: [sizeO, sizeO + heightOverSet],
-    }),
-    width: sizeO,
-  };
-
-  const boxOverlayStyles = {
-    height: overlayAnim.interpolate({
-      inputRange: [0, 1],
-      outputRange: ['0%', '100%'],
-    }),
-    width: overlayAnim.interpolate({
-      inputRange: [0, 1],
-      outputRange: ['0%', '100%'],
-    }),
-  };
-
-  useEffect(() => {
-    const { active } = pinBoxStates;
-    if (active) {
-      Animated.timing(checkboxAnim, {
-        toValue: 1,
-        duration: 50,
-        easing: Easing.cubic,
-        useNativeDriver: false,
-      }).start();
-    }
-  }, [checkboxAnim, pinBoxStates]);
-
-  const setValue = (value = 0) => {
-    setPinBoxStates((prevState) => ({ ...prevState, value }));
-    Animated.timing(checkboxAnim, {
-      toValue: 0,
-      duration: 25,
-      easing: Easing.cubic,
-      useNativeDriver: false
-    }).start();
-  }
-
-  const setValueWO = (value) => {
-    setPinBoxStates((prevState) => ({ ...prevState, value }));
-  }
-
-  const activate = () => {
-    setPinBoxStates((prevState) => ({ ...prevState, active: true }));
-    Animated.timing(checkboxAnim, {
-      toValue: 1,
-      duration: 50,
-      easing: Easing.cubic,
-      useNativeDriver: false
-    }).start();
-  }
-
-  const deactivate = (callBack = undefined) => {
-    setPinBoxStates((prevState) => ({ ...prevState, active: false }));
-    Animated.timing(checkboxAnim, {
-      toValue: 0,
-      duration: 50,
-      easing: Easing.cubic,
-      useNativeDriver: false
-    }).start(callBack);
-  }
-
-  const deactivateWO = (callBack = undefined) => {
-    setPinBoxStates((prevState) => ({ ...prevState, active: false }));
-    Animated.timing(checkboxAnim, {
-      toValue: 0,
-      duration: 5,
-      easing: Easing.cubic,
-      useNativeDriver: false
-    }).start(callBack);
-  }
-
-  const removeValue = (callBack = undefined) => {
-    setPinBoxStates((prevState) => ({ ...prevState, hasValue: false }));
-    setTimeout(() => {
-      Animated.timing(overlayAnim, {
-        toValue: 0,
-        duration: 50,
-        easing: Easing.cubic,
-        useNativeDriver: false
-      }).start();
-    }, 250);
-    typeof callBack === 'function' && callBack();
-  }
-
-  const fixValue = (callBack = undefined) => {
-    setPinBoxStates((prevState) => ({ ...prevState, hasValue: true }));
-    setTimeout(() => {
-      Animated.timing(overlayAnim, {
-        toValue: 1,
-        duration: 50,
-        easing: Easing.cubic,
-        useNativeDriver: false
-      }).start();
-    }, 250);
-    typeof callBack === 'function' && callBack();
-  }
-
-  useImperativeHandle(forwardingRef, () => {
-    return { setValue, setValueWO, activate, deactivate, deactivateWO, removeValue, fixValue }
-  })
-
-  const { value, MaxHeight, MaxWidth } = pinBoxStates;
-
-  return (
-    <>
-      <View
-        style={[
-          PinInputStyles.PinBoxContainer,
-          { height: MaxHeight, width: MaxWidth },
-        ]}>
-        <Animated.View
-          style={[
-            boxStyles,
-            PinInputStyles.PinBox,
-            PinInputStyles.inputShadow,
-          ]}
-        >
-          {pinBoxStates.hasValue && (
-            <Animated.View
-              style={[boxOverlayStyles, PinInputStyles.PinBoxOverlay]}>
-              <View
-                style={[
-                  PinInputStyles.overlayDot,
-                  PinInputStyles.inputShadow,
-                ]}
-              />
-            </Animated.View>
-          )}
-          <Text style={[{ fontWeight: 'bold', fontSize: 17 }]}>
-            {value}
-          </Text>
-        </Animated.View>
-        <Text>{value}</Text>
-        <Text>{JSON.stringify(pinBoxStates.hasValue)}</Text>
-      </View>
-    </>
-  );
-
-});
+import { PinCellCurvy } from './PinCellCurvy';
 
 export const PinInputCurvy = forwardRef((PinInputCurvyProps, forwardingRef) => {
   const { pinLength, size, onTap, onPinInputComplete, pinLengthZero } = PinInputCurvyProps;
@@ -291,7 +131,7 @@ export const PinInputCurvy = forwardRef((PinInputCurvyProps, forwardingRef) => {
           {[...Array(pinLength).keys()]?.map(i => {
             const eleRef = createRef();
             BoxRefs.push(eleRef);
-            return <PinBox key={i} size={size || 60} ref={eleRef} />;
+            return <PinCellCurvy key={i} size={size || 60} ref={eleRef} />;
           })}
         </Animated.View>
       </TouchableWithoutFeedback>
